Move small utils into own folder
