+++ src/utils/colors.scss
@@ -1,5 +1,5 @@
 @use "sass:color";
-@import "./variables";
+@import "../variables";
 
 :export {
     nordicBlue: $nordic-blue;
